requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
email-validator>=2.0.0
